api changes

+++ backend/src/api/controllers/userController.js
@@ -93,7 +93,7 @@ const signup = async (req, res) => {
     // Save the new user
     const userSavedData = await newUser.save();
 
-   
+
     // Send a verification email
     sendVerifyEmail(firstName, email, newUser._id);
 
@@ -174,7 +174,7 @@ const update = async (req, res) => {
     if (req.file) {
       const file = req.file;
       const filePath = `/uploads/${file.filename}`;
-      req.body.profilePicture = filePath; 
+      req.body.profilePicture = filePath;
     }
 
     delete req.body.password;
@@ -185,29 +185,29 @@ const update = async (req, res) => {
     );
 
     // Generate new token
-const token = jwt.sign(
-  { 
-    id: updatedContact._id, 
-    name: updatedContact.firstName, 
-    role: updatedContact.role 
-  },
-  process.env.JWT_SECRET,
-  { expiresIn: 86400 } 
-);
+    const token = jwt.sign(
+      {
+        id: updatedContact._id,
+        name: updatedContact.firstName,
+        role: updatedContact.role
+      },
+      process.env.JWT_SECRET,
+      { expiresIn: 86400 }
+    );
 
-res.status(200).json({
-  _id: updatedContact._id,
-  firstName: updatedContact.firstName,
-  lastName: updatedContact.lastName,
-  email: updatedContact.email,
-  role: updatedContact.role,
-  companyName: updatedContact.companyName,
-  roles: updatedContact.roles,
-  isEmailVerified: updatedContact.isEmailVerified,
-  createdAt: updatedContact.createdAt,
-  updatedAt: updatedContact.updatedAt,
-  accessToken: token
-});
+    res.status(200).json({
+      _id: updatedContact._id,
+      firstName: updatedContact.firstName,
+      lastName: updatedContact.lastName,
+      email: updatedContact.email,
+      role: updatedContact.role,
+      companyName: updatedContact.companyName,
+      roles: updatedContact.roles,
+      isEmailVerified: updatedContact.isEmailVerified,
+      createdAt: updatedContact.createdAt,
+      updatedAt: updatedContact.updatedAt,
+      accessToken: token
+    });
   } catch (error) {
     return res.status(500).json({ message: error.message, status: 500 });
   }
@@ -215,10 +215,17 @@ res.status(200).json({
 
 const deleteById = async (req, res) => {
   try {
+    if (!req.query.id) {
+      return res.status(400).json({ message: "id is required" })
+    }
+    const exist = await userModel.findById(req.query.id).select("name");
+    if (!exist) {
+      return res.status(404).json({ message: "User not found" });
+    }
     const result = await userModel.findByIdAndDelete({
       _id: req.query._id || req.query.id,
     });
-    res.status(200).json({ result });
+    res.status(200).json({ message:"User deleted successfully." });
   } catch (error) {
     return res.status(500).json({ message: error.message, status: 500 });
   }
@@ -253,7 +260,7 @@ const findAll = async (req, res) => {
   }
 };
 
-const reset_password = async (req, res) => {
+const resetPassword = async (req, res) => {
   try {
     const token = req.body.token;
     const tokenData = await userModel.findOne({ token: token });
@@ -282,7 +289,7 @@ const reset_password = async (req, res) => {
 
 const sendResetPasswordMail = async (name, email, token) => {
   try {
-    const html = `<p> Hi ${name}, please copy the link <a href="https://abc.com/reset-password?token=${token}"> reset your password </a>.</p>`;
+    const html = `<p> Hi ${name}, please copy the link <a href="https://amplifier.hgsingalong.com/resetPassword?token=${token}"> reset your password </a>.</p>`;
     await sendEmail(email, "For Reset password", html);
   } catch (error) {
     return res.status(500).json({ message: error.message, status: 500 });
@@ -317,7 +324,7 @@ const forgotPassword = async (req, res) => {
 
 const verifymail = async (req, res) => {
   const id = req.query.id;
-  
+
   try {
 
     const user = await userModel.findOne({ _id: id });
@@ -334,7 +341,7 @@ const verifymail = async (req, res) => {
     const verifiedMail = await userModel.updateOne(
       { _id: id }, // Correct query object
       { $set: { isEmailVerified: true } },
-      
+
     );
     return res.status(200).json({ message: 'Email successfully verified', status: 200 });
   } catch (error) {
@@ -469,7 +476,7 @@ module.exports = {
   deleteById,
   findById,
   findAll,
-  reset_password,
+  resetPassword,
   forgotPassword,
   verifymail,
   uploadUserExcel,
